requirements.txt
astroid==2.3.3
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
choice==0.1
colorama==0.4.3
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
emoji==0.5.4
entrypoints==0.3
future==0.18.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.1
python-telegram-bot==12.2.0
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
Send2Trash==1.5.0
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
clarifai
